backend/requirements.txt
annotated-types==0.7.0
anyio==4.8.0
asgiref==3.7.2
attrs==25.3.0
beautifulsoup4==4.13.3
certifi==2025.1.31
cffi==1.17.1
charset-normalizer==3.4.1
click==8.1.8
colorama==0.4.6
cryptography==45.0.2
defusedxml==0.7.1
distro==1.9.0
dj-rest-auth==7.0.1
Django==4.2.16
django-allauth==0.61.1
django-cors-headers==4.6.0
djangorestframework==3.15.2
djangorestframework_simplejwt==5.5.0
et_xmlfile==2.0.0
exceptiongroup==1.2.2
gTTS==2.5.4
h11==0.14.0
httpcore==1.0.7
httpx==0.28.1
idna==3.10
jiter==0.8.2
joblib==1.4.2
numpy==2.0.2
oauthlib==3.2.2
openai==1.61.1
openpyxl==3.1.5
outcome==1.3.0.post0
pandas==2.2.3
Pillow==10.0.1
pycparser==2.22
pydantic==2.10.6
pydantic_core==2.27.2
PyJWT==2.9.0
PySocks==1.7.1
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
python-environ==0.4.54
python3-openid==3.2.0
pytz==2023.3.post1
requests==2.32.3
requests-oauthlib==2.0.0
scikit-learn==1.6.1
scipy==1.13.1
selenium==4.30.0
six==1.17.0
sniffio==1.3.1
sortedcontainers==2.4.0
soupsieve==2.6
sqlparse==0.4.4
threadpoolctl==3.6.0
tqdm==4.67.1
trio==0.29.0
trio-websocket==0.12.2
typing-inspection==0.4.0
typing_extensions==4.12.2
tzdata==2023.3
urllib3==2.3.0
websocket-client==1.8.0
Wikipedia-API==0.8.1
wsproto==1.2.0
